ProcessController: Add casts to appease GCC2.

+++ src/apps/processcontroller/ProcessController.cpp
@@ -150,9 +150,9 @@ ProcessController::ComposeSize(float maxWidth, float maxHeight)
 
 	// Scale, and be at least as wide as tall
 	if (maxHeight > 16)
-		width *= (maxHeight / 16);
+		width *= (int)(maxHeight / 16);
 	if (width < maxHeight)
-		width = maxHeight;
+		width = (int)maxHeight;
 
 	// Damn, you got a lot of CPU
 	if (width > maxWidth)
